Fixed responsiveness issues and moved code out of a file
