+++ src/Layouts/Main/YouTubeVideo/Components/SubmitButton.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Button } from '@mui/material';
 import PropTypes from 'prop-types';
-import { Colors } from '../../../../../constants/Colors';
+import { Colors } from '../../../../constants/Colors';
 
 export default function SubmitButton({ handleSubmit }) {
 	return (
